Move package install/uninstall functionality to the core project.
