+++ SharpAdbClient.Extensions.Tests/AdbHelperTests.cs
@@ -1,12 +1,7 @@
-﻿using SharpAdbClient.IO;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SharpAdbClient.Tests;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SharpAdbClient.Extensions.Tests
 {
