+++ SharpAdbClient.Extensions.Tests/SharpAdbClient.Extensions.Tests.csproj
@@ -70,7 +70,6 @@
     <Compile Include="DummyDevice.cs" />
     <Compile Include="FileEntryTests.cs" />
     <Compile Include="FileSystemTests.cs" />
-    <Compile Include="LinuxPathTests.cs" />
     <Compile Include="ListingServiceReceiverTests.cs" />
     <Compile Include="ListingServiceTests.cs" />
     <Compile Include="PackageManagerTests.cs" />
