+++ SharpAdbClient.Extensions/FileEntry.cs
@@ -1,11 +1,10 @@
-﻿using System;
+﻿using SharpAdbClient.DeviceCommands;
+using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
-using System.IO;
-using SharpAdbClient.IO;
-using MoreLinq;
 
 namespace SharpAdbClient
 {
