+++ SharpAdbClient.Extensions/SharpAdbClient.Extensions.csproj
@@ -64,7 +64,6 @@
     <Compile Include="BusyBox.cs" />
     <Compile Include="Device.cs" />
     <Compile Include="DeviceEventArgs.cs" />
-    <Compile Include="Exceptions\PackageInstallationException.cs" />
     <Compile Include="FileEntry.cs" />
     <Compile Include="FileListingService.cs" />
     <Compile Include="FilePermission.cs" />
@@ -74,7 +73,6 @@
     <Compile Include="IDevice.cs" />
     <Compile Include="IFileListingService.cs" />
     <Compile Include="IFileSystem.cs" />
-    <Compile Include="IO\LinuxPath.cs" />
     <Compile Include="MountPoint.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Receivers\BatteryReceiver.cs" />
@@ -100,6 +98,10 @@
       <Project>{65473257-e70f-410b-9269-d0c0f771ea87}</Project>
       <Name>Managed.Adb</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Exceptions\" />
+    <Folder Include="IO\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
