+++ SharpAdbClient/DeviceCommands/InstallReceiver.cs
@@ -2,16 +2,13 @@
 // Copyright (c) The Android Open Source Project, Ryan Conrad, Quamotion. All rights reserved.
 // </copyright>
 
-namespace SharpAdbClient
+namespace SharpAdbClient.DeviceCommands
 {
-    using System;
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
     using System.Text.RegularExpressions;
 
     /// <summary>
-    ///
+    /// 
     /// </summary>
     public class InstallReceiver : MultiLineReceiver
     {
