+++ SharpAdbClient/DeviceCommands/LinuxPath.cs
@@ -2,7 +2,7 @@
 // Copyright (c) The Android Open Source Project, Ryan Conrad, Quamotion. All rights reserved.
 // </copyright>
 
-namespace SharpAdbClient.IO
+namespace SharpAdbClient.DeviceCommands
 {
     using System;
     using System.Collections.Generic;
